+++ src/triangle.h
@@ -1,1 +1,1 @@
-double triangle_herons_formula(double a, double b, double c);
+long double triangle_herons_formula(long double a, long double b, long double c);
